moving getFilter to utils
